@@ -32,23 +32,24 @@
   "typings": "./ng2-dnd.d.ts",
   "homepage": "https://github.com/akserg/ng2-dnd#readme",
   "peerDependencies": {
-    "@angular/core": "^4.0.0 || ^5.0.0",
-    "@angular/forms": "^4.0.0 || ^5.0.0"
+    "@angular/core": "^8.0.0 || ^9.0.0",
+    "@angular/forms": "^8.0.0 || ^9.0.0"
   },
   "devDependencies": {
-    "@angular/common": "^5.0.0",
-    "@angular/compiler": "^5.0.0",
-    "@angular/compiler-cli": "^5.0.0",
-    "@angular/core": "^5.0.0",
-    "@angular/forms": "^5.0.0",
-    "@angular/platform-browser": "^5.0.0",
-    "@angular/platform-browser-dynamic": "^5.0.0",
-    "@angular/platform-server": "^5.0.0",
+    "@angular/animations": "^9.0.1",
+    "@angular/common": "^9.0.0",
+    "@angular/compiler": "^9.0.0",
+    "@angular/compiler-cli": "^9.0.0",
+    "@angular/core": "^9.0.0",
+    "@angular/forms": "^9.0.0",
+    "@angular/platform-browser": "^9.0.0",
+    "@angular/platform-browser-dynamic": "^9.0.0",
+    "@angular/platform-server": "^9.0.0",
     "@types/hammerjs": "^2.0.34",
     "@types/jasmine": "^2.5.51",
     "@types/node": "^8.0.53",
     "awesome-typescript-loader": "^3.1.3",
-    "codelyzer": "^4.0.1",
+    "codelyzer": "^5.1.2",
     "core-js": "^2.4.1",
     "istanbul-instrumenter-loader": "^3.0.0",
     "jasmine-core": "^2.6.2",
@@ -62,18 +63,19 @@
     "karma-sourcemap-loader": "^0.3.7",
     "karma-webpack": "^2.0.3",
     "loader-utils": "^1.1.0",
-    "ng-packagr": "^2.0.0-rc.0",
+    "ng-packagr": "^9.0.1",
     "reflect-metadata": "^0.1.10",
-    "rxjs": "^5.4.0",
+    "rxjs": "^6.5.3",
     "source-map-loader": "^0.2.1",
     "ts-helpers": "1.1.2",
+    "tslib": "^1.10.0",
     "tslint": "^5.4.3",
     "tslint-loader": "^3.5.3",
-    "typescript": "^2.3.4",
+    "typescript": "^3.7.4",
     "webpack": "^3.6.0",
-    "zone.js": "^0.8.12"
+    "zone.js": "~0.10.2"
   },
   "engines": {
-    "npm": ">4.0.0"
+    "npm": ">6.0.0"
   }
 }
